Move copyFile to /lib
