+++ lib/copyFile.js
@@ -1,5 +1,5 @@
-import { getGraphClient, getAccessToken } from "./lib/msAuth.js";
-import { getFileIdByName } from "./lib/oneDrive.js";
+import { getGraphClient, getAccessToken } from "./msAuth.js";
+import { getFileIdByName } from "./oneDrive.js";
 
 // Function to check if a file with the same name exists
 const checkFileExists = async (userId, folderId, fileName) => {
