+++ lib/worksheetProcessing.js
@@ -8,7 +8,7 @@ import {
 } from "./utils.js";
 import { getFileIdByName } from "./oneDrive.js";
 import { clearWorksheetRange } from "./clearWorksheet.js";
-import { copyFileInOneDrive } from "../copyFile.js";
+import { copyFileInOneDrive } from "./copyFile.js";
 // import { copyFileInOneDrive } from "../copyFile.js";
 // import { clearWorksheetRange } from "../clearWorksheet.js";
 import fs from "fs";
